+++ repository/build.gradle
@@ -76,6 +76,7 @@ dependencies {
 
     implementation project(path: ':data')
     implementation project(path: ':common')
+    implementation project(path: ':preference')
     //app auth now added as source code
     //implementation "net.openid:appauth:$versions.appauth"
     implementation project(path: ':appauth')
