+++ repository/src/main/java/ca/bc/gov/repository/OnBoardingRepository.kt
@@ -1,7 +1,7 @@
 package ca.bc.gov.repository
 
 import ca.bc.gov.common.BuildConfig
-import ca.bc.gov.data.datasource.local.preference.EncryptedPreferenceStorage
+import ca.bc.gov.preference.EncryptedPreferenceStorage
 import javax.inject.Inject
 
 /**
