+++ repository/src/main/java/ca/bc/gov/repository/di/AnalyticsModule.kt
@@ -1,6 +1,6 @@
 package ca.bc.gov.repository.di
 
-import ca.bc.gov.data.datasource.local.preference.EncryptedPreferenceStorage
+import ca.bc.gov.preference.EncryptedPreferenceStorage
 import ca.bc.gov.repository.analytics.AnalyticsRepository
 import dagger.Module
 import dagger.Provides
