+++ scripts/versions.gradle
@@ -68,6 +68,8 @@ versions.biometric = "1.2.0-alpha04" //1.2.0-alpha05 requires 33+
 
 versions.appauth = "0.11.0"
 
+versions.coroutines = "1.6.4"
+
 versions.junit = "4.13.2"
 versions.ext_junit = "1.1.5"
 versions.espresso = "3.5.1"
