+++ settings.gradle
@@ -2,5 +2,6 @@ rootProject.name = "My-Health"
 include ':app'
 include ':data'
 include ':repository'
-include ':common'
+include ':preference'
 include ':appauth'
+include ':common'
